+++ js/bday_misc.js
@@ -35,6 +35,39 @@ var BDAY_toggleSub = function(cbox, uid) {
 };
 
 /**
+ * Toggle a user's subscription to receive birthday cards.
+ *
+ * @param   object  cbox    Subscription yes/no checkbox
+ * @param   integer uid     User ID
+ * @return  boolean     False
+ */
+var BDAY_toggleCards = function(cbox, uid) {
+    var oldval = cbox.checked ? 0 : 1;
+    var dataS = {
+        "action" : "toggleCards",
+        "uid": uid,
+        "oldval": oldval,
+    };
+    data = $.param(dataS);
+    $.ajax({
+        type: "POST",
+        dataType: "json",
+        url: glfusionSiteUrl + "/birthdays/ajax.php",
+        data: data,
+        success: function(result) {
+            cbox.checked = result.newval == 1 ? true : false;
+            try {
+                $.UIkit.notify("<i class='uk-icon-check'></i>&nbsp;" + result.statusMessage, {timeout: 1000,pos:'top-center'});
+            }
+            catch(err) {
+                alert(result.statusMessage);
+            }
+        }
+    });
+    return false;
+};
+
+/**
  * Update the days selection based on the supplied month.
  *
  * @param   integer value   Month number