+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog for the Birthdays plugin
 
+## Version 1.2.0
+Release TBD
+- Require glFusion 2.0.1
+- Add customizable template header for notifications.
+- Add admin access to toggle user's preference to receive cards.
+- Fix parameters passed to `PLG_sendSubscriptionNotification`.
+
 ## Version 1.1.1
 Release 2021-12-16
 - Add group-based access to view and cards as an override to features.